requirements.txt
arrow==0.12.1
backports-abc==0.5
certifi==2018.10.15
chardet==3.0.4
discord-webhook==0.4.1
idna==2.7
lxml==4.4.1
praw==6.0.0
prawcore==1.0.0
psaw==0.0.7
python-dateutil==2.7.5
PyYAML==3.13
requests==2.20.0
scraper==0.1.0
shreddit==6.0.7
six==1.11.0
tornado==5.1.1
update-checker==0.16
urllib3==1.24.2
